requirements.txt
Django==1.10.1
djangorestframework==3.4.0
django-rest-swagger==0.3.10
opencv-python==4.2.0.34
pyzbar==0.1.8
gunicorn==19.6.0
sentry-sdk==0.14.3
django-log-request-id==1.3.2
python-json-logger==0.1.4

